requirements.txt
# Storypark Scraper – Python dependencies
# Install with: pip install -r requirements.txt

# Web scraping
playwright>=1.44.0

# Facial recognition
face_recognition>=1.3.0

# Image processing & EXIF manipulation
Pillow>=10.0.0
piexif>=1.1.3

# Google Photos upload
google-auth>=2.29.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
requests>=2.31.0
